revising to load data correctly

+++ rfmix_reader/_fb_reader.h
@@ -7,4 +7,4 @@
 
 void read_fb_chunk(float *buff, uint64_t nrows, uint64_t ncols,
 		   uint64_t row_start, uint64_t col_start, uint64_t row_end,
-		   uint64_t col_end, float *out, uint64_t *strides);
+		   uint64_t col_end, uint8_t *out, uint64_t *strides);
